requirements.txt
jinja2==2.10
boto3==1.9.79
requests2==2.16.0